12/31/2020
Finished up the code that allows the program to get software variables from a file and put them into the correct variable. Also fixed a bug that made it to where when the user pressed the back button on the registration page the welcome page would open in a new scene.