1/11/2021
Finished the main parts of the login and registration pages plus on top of that the database controller. The login page now gets and checks the users input up against the database. The registration page now puts the inputed data into the database. Need to add a link to the registration page from the login page. Need to check if the username has been already put into the database on the registration page. I'll finish that in the next push