1/14/2021
Added the username check on the registration page. Now if the user goes to the page and the database returns a username for that username then it will tell the user to use a different username. Also added a link on the login and registration that allows the user to register if need be or sign up if need be.  Also fixed a bug with the SeenWelcomePage boolean writing method where if you went back to the welcome page and went to either login or registration it would rewrite the boolean into the SystemInformation.info file. Also started to put the users host ip, username and password into the database. I will work soon on encrypting that data with probably a key that will stay on the users computer and encrypt and decrypt users data. Also I will start working on the main screen but I need to do some planning for it first like the command groups, how to make scrolling pages and how to execute commands and what the layout will look like.