3/8/21
I believe (I haven't worked on any projects in a while due to school) that I was trying to get a simple application installer created and it seems I have maybe about less than half of it done. I'm working on a error popup screen but keep getting an null pointer exception when putting text into it. It's annoying but I'll figure it out in a little while. I plan to finish it up and then continue working on the main application.